Add tvim, a wrapper script to start Vim in a modern terminal.
I'm a big fan of gvim. gvim is fast, has a better render quality than
most classic terminals and good desktop integration. But on unixlike
systems gvim is currently tied to X11. While there is some work done
to get gvim working on Wayland, that effort is progressing slowly.
There're a lot of corner cases like +clientserver to solve and it will
likely take years to get ready for primetime and become available in
all common distros.

So it's time bite the bullet and build a "terminal gvim", in short tvim.
This script starts a modern terminal with all the nice features like a
very high rendering quality or colored undercurls and Vim in it. If Vim
is correctly configured, it will behave more or less the same as gvim.