Introduce an artifical delay of 25 ms before starting vim.
This is a an hack working around the alacritty executing the command
before the window was created. Starting vim in an uncreated windows
leads to vim determining a wrong window size which in turn leads to
rendering problems.

Apply the same to gnome-terminal, even if I haven't seen the problem
there.